Removes extra gunk.
